requirements.txt
PyPDF2
pdfplumber
python-dateutil
openai
python-dotenv
pypdf
streamlit